Fix: product 관련 파일 삭제, search와 list 합침
