+++ controllers/index.js
@@ -1,21 +1,17 @@
-import categoryController from './categoryController';
 import productController from './productController';
 import commentController from './commentController';
 import cartController from './cartController';
 import likeController from './likeController';
 import loginController from './loginController';
 import signupController from './signupController';
-import searchController from './searchController';
 import listController from './listController';
 
 export {
-  categoryController,
   productController,
   commentController,
   cartController,
   likeController,
   loginController,
   signupController,
-  searchController,
   listController,
 };
