+++ models/index.js
@@ -1,21 +1,17 @@
-import categoryDao from './categoryDao';
 import productDao from './productDao';
 import commentDao from './commentDao';
 import cartDao from './cartDao';
 import likeDao from './likeDao';
 import loginDao from './loginDao';
 import signupDao from './signupDao';
-import searchDao from './searchDao';
 import listDao from './listDao';
 
 export {
-  categoryDao,
   productDao,
   commentDao,
   cartDao,
   likeDao,
   loginDao,
   signupDao,
-  searchDao,
   listDao,
 };
