+++ routes/listRouter.js
@@ -5,5 +5,6 @@ const router = express.Router();
 
 router.get('/:depth/:id', listController.getProductsForEachCategory);
 router.get('/:sort', listController.getCategorizedProducts);
+router.get('/', listController.searchProducts);
 
 export default router;
