+++ services/index.js
@@ -1,21 +1,17 @@
-import categoryServices from './categoryServices';
 import productServices from './productServices';
 import commentServices from './commentServices';
 import cartServices from './cartServices';
 import likeServices from './likeServices';
 import loginServices from './loginServices';
 import signupServices from './signupServices';
-import searchServices from './searchServices';
 import listServices from './listServices';
 
 export {
-  categoryServices,
   productServices,
   commentServices,
   cartServices,
   likeServices,
   loginServices,
   signupServices,
-  searchServices,
   listServices,
 };
